+++ tests/test_table_store_elastic.py
@@ -1,0 +1,19 @@
+import pytest
+
+from datapipe.store.elastic import ElasticStore
+from datapipe.store.tests.abstract import AbstractBaseStoreTests
+from datapipe.types import DataSchema
+
+
+class TestElasticStore(AbstractBaseStoreTests):
+    @pytest.fixture
+    def store_maker(self, elastic_conn):
+        def make_elastic_store(data_schema: DataSchema):
+            return ElasticStore(
+                elastic_conn["index"],
+                data_schema,
+                elastic_conn["es_kwargs"],
+                {},
+            )
+
+        return make_elastic_store